+++ include/WebServerException/WebServerException.hpp
